+++ src/variables.mjs
@@ -26,7 +26,7 @@ export { default as to_int } from './variables/to_int.mjs';
 export { default as to_integer } from './variables/to_integer.mjs';
 export { default as isset } from './variables/isset.mjs';
 export { default as to_string } from './variables/to_string.mjs';
-export { default as print_r } from './variables/print_r.mjs';
+export { default as print } from './variables/print.mjs';
 export { default as var_dump } from './variables/var_dump.mjs';
 export { default as serialize } from './variables/serialize.mjs';
 export { default as settype } from './variables/settype.mjs';