+++ tests/variables.global.test.mjs
@@ -30,7 +30,7 @@ test('variables.global', () => {
     expect(to_integer).not.toBeUndefined();
     expect(isset).not.toBeUndefined();
     expect(to_string).not.toBeUndefined();
-    expect(print_r).not.toBeUndefined();
+    expect(print).not.toBeUndefined();
     expect(var_dump).not.toBeUndefined();
     expect(serialize).not.toBeUndefined();
     expect(settype).not.toBeUndefined();